requirements.txt
futu-api>=8.0.0
pandas>=2.1.0
